Insert zero-width non-breaking space instead
After more testing, it is a better UX to insert
a zero-width space because:
1. When a space is inserted to place the cursor
afterward, the user occassionally hits backspace
to remove it, and doing so places the cursor
back within the token (at least in Chrome 66).
If we use zero-width non-breaking spaces,
the user can simply add a space when desired
and a reduced chance of accidentally re-placing
the cursor within the token immediately
after inserting it.
2. TinyMCE seems to handle ZWNBSPs well at
token boundaries when moving the cursor
with the keyboard.